agrege el getion de personal

+++ turnix-backend/src/main/java/com/turnix/turnix_backend/repository/ProfesionalRepository.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProfesionalRepository extends JpaRepository<Profesional, Long> {
+
+public interface ProfesionalRepository extends JpaRepository<Profesional, Integer> {
     List<Profesional> findByNegocio_Id(Integer negocioId);
 }
